Various improvements: - UMTX1 implementation working again. - Various attempts to fix deadlocking in UMTX1 implementation. Not sure completely if effective. - Added Executor helper class to run the exploit in a separate thread and wait. - Fixed an issue with Memory dumper not printing correct data. - Added logging config UI to UMTX1. - Fixed deadlock in UMTX2 implementation. Picked latest changes from webkit repo.

+++ xploit/src/main/java/org/ps5jb/client/payloads/umtx/impl1/KernelStackMarkerMethod.java
@@ -10,10 +10,10 @@ public final class KernelStackMarkerMethod {
     // TODO: Not able to get IOCTL method to work...Maybe need to play with MAX_RECLAIM_SYSTEM_CALLS.
     public static final KernelStackMarkerMethod IOCTL = new KernelStackMarkerMethod(2, 128, 32, 250);
 
-    public static final KernelStackMarkerMethod SCHED_YIELD = new KernelStackMarkerMethod(3, 0,2, 100);
+    public static final KernelStackMarkerMethod SCHED_YIELD = new KernelStackMarkerMethod(3, 0, 2, 100);
 
     private int id;
-    private int markerSize;
+    private long markerSize;
     private int searchLoopInvocations;
     private long kernelStackWaitPeriod;
 
@@ -28,7 +28,7 @@ public final class KernelStackMarkerMethod {
         return id;
     }
 
-    public int getMarkerSize() {
+    public long getMarkerSize() {
         return markerSize;
     }
